Reorganize importing QKeras (part 2)
